logger utility function + screentime now part of timer + prettyPrint now located in helpers
